add: strlen to get full command line checked

+++ srcs/signal.c
@@ -4,8 +4,10 @@ void	sigquit_handler(int signal)
 {
 	if (signal == SIGQUIT || signal == SIGINT)
 	{
-		// printf("Signal received\n");
-		kill(child_process, signal);
+		if (child_process > -1)
+			kill(child_process, signal);
+		if (signal == SIGINT)
+			write(0, "\n", 1);
 	}
 	else
 		printf("Received another signal\n");
